cambios para que funcione el pago con tarjetas

+++ src/main/java/upv/etsinf/cognispatium/repository/JPATarjetaDao.java
@@ -36,7 +36,7 @@ public class JPATarjetaDao implements TarjetaDao {
 	}
 
     @Override
-    public Tarjeta getTarjetaById(Long tarjetaId) {
+    public Tarjeta getTarjetaById(int tarjetaId) {
         return em.find(Tarjeta.class,tarjetaId);
     }
