+++ src/main/java/upv/etsinf/cognispatium/service/SimpleTarjetaManager.java
@@ -34,7 +34,7 @@ public class SimpleTarjetaManager implements Serializable {
 		tarjetaDao.saveTarjeta(tarjeta);
 	}
 
-	public Tarjeta getTarjetaByID(Long tarjetaId) {
+	public Tarjeta getTarjetaByID(int tarjetaId) {
 	    return tarjetaDao.getTarjetaById(tarjetaId);
 	}
 }
